requirements.txt
kafka-python==2.0.1
psycopg2==2.8.5
requests==2.24.0